Clean up duplication
